implemented first combat
implemented punching:
all enemies in range will be scanned, the one closest to the direction you are facing will be chosen and your character snaps to it (meaning it turns towards it). then, based on the attack angle, all enemies in a cone in front of the character will be registered and "damaged" (currently console output).
try it out with the 2 cubes in the scene, all cobat parameters can be changed in the PlayerGirl prefab in the prefabs folder.